Rename CustomSerializer -> CustomEventSerializer
