+++ README.md
@@ -50,7 +50,7 @@ A number fo the documents below are still a work in progress and would be added 
 
 * [Simple Consumer](./samples/SimpleConsumer)
 * [Simple Publisher](./samples/SimplePublisher)
-* [Build a custom serializer](./samples/CustomSerializer)
+* [Build a custom event serializer](./samples/CustomEventSerializer)
 * [Consume multiple events in one consumer](./samples/MultiEventsConsumer)
 * [In memory background processing](./samples/InMemoryBackgroundProcessing)
