+++ Tingle.EventBus.sln
@@ -48,7 +48,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "samples", "samples", "{62F6
 		samples\Directory.Build.props = samples\Directory.Build.props
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomSerializer", "samples\CustomSerializer\CustomSerializer.csproj", "{2C55FABC-8C94-4104-BE05-42A477D2AD9E}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CustomEventSerializer", "samples\CustomEventSerializer\CustomEventSerializer.csproj", "{2C55FABC-8C94-4104-BE05-42A477D2AD9E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "InMemoryBackgroundProcessing", "samples\InMemoryBackgroundProcessing\InMemoryBackgroundProcessing.csproj", "{C9293277-90BA-4F1A-BEA7-85CE41103B8D}"
 EndProject
@@ -108,6 +108,22 @@ Global
 		{1BCD0C98-D061-4143-B00F-0031FBCD5E5E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1BCD0C98-D061-4143-B00F-0031FBCD5E5E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1BCD0C98-D061-4143-B00F-0031FBCD5E5E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B52EE5B6-D86E-4415-8688-E835220AD219}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B52EE5B6-D86E-4415-8688-E835220AD219}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B52EE5B6-D86E-4415-8688-E835220AD219}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B52EE5B6-D86E-4415-8688-E835220AD219}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{2C55FABC-8C94-4104-BE05-42A477D2AD9E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{2C55FABC-8C94-4104-BE05-42A477D2AD9E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2C55FABC-8C94-4104-BE05-42A477D2AD9E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -128,22 +144,6 @@ Global
 		{2FEC41FE-188B-4B02-943B-0FEAB0E1FA39}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2FEC41FE-188B-4B02-943B-0FEAB0E1FA39}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2FEC41FE-188B-4B02-943B-0FEAB0E1FA39}.Release|Any CPU.Build.0 = Release|Any CPU
-		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6536B1EC-D297-473F-9C18-DD42304CB9FE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B52EE5B6-D86E-4415-8688-E835220AD219}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B52EE5B6-D86E-4415-8688-E835220AD219}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B52EE5B6-D86E-4415-8688-E835220AD219}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B52EE5B6-D86E-4415-8688-E835220AD219}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -160,15 +160,15 @@ Global
 		{07C3232C-687A-4C85-9A67-9BEF854BB91F} = {3647730E-193B-45E9-8ECE-DDA7494C9030}
 		{2664F6EA-94DB-4E14-A5B3-43FBCCF472E3} = {3647730E-193B-45E9-8ECE-DDA7494C9030}
 		{1BCD0C98-D061-4143-B00F-0031FBCD5E5E} = {3647730E-193B-45E9-8ECE-DDA7494C9030}
+		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
+		{6536B1EC-D297-473F-9C18-DD42304CB9FE} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
+		{B52EE5B6-D86E-4415-8688-E835220AD219} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
+		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
 		{2C55FABC-8C94-4104-BE05-42A477D2AD9E} = {62F603F3-FF36-4E36-AC0C-08D1883525BE}
 		{C9293277-90BA-4F1A-BEA7-85CE41103B8D} = {62F603F3-FF36-4E36-AC0C-08D1883525BE}
 		{E9D6F25A-0586-4409-A3EB-A72B6E89A71C} = {62F603F3-FF36-4E36-AC0C-08D1883525BE}
 		{ED1E8BA4-9538-42AB-9759-4FC4233A36B1} = {62F603F3-FF36-4E36-AC0C-08D1883525BE}
 		{2FEC41FE-188B-4B02-943B-0FEAB0E1FA39} = {62F603F3-FF36-4E36-AC0C-08D1883525BE}
-		{108968DB-1C48-4B53-9F23-CE7C5BAA88B4} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
-		{6536B1EC-D297-473F-9C18-DD42304CB9FE} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
-		{B52EE5B6-D86E-4415-8688-E835220AD219} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
-		{C2E8CDC9-E607-4DDF-9796-98EF9CAFC65D} = {BDD324B6-9EFC-49A3-9CF6-6CE494446C4B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9E5CC03E-DDE6-4455-9D5D-03617DE5DBD8}
