+++ samples/CustomEventSerializer/CustomEventSerializer.csproj
