+++ samples/CustomEventSerializer/Models/AzureDevOpsEventResource.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace CustomSerializer.Models
+namespace CustomEventSerializer.Models
 {
     public class AzureDevOpsEventResource
     {
