+++ samples/CustomEventSerializer/Models/AzureDevOpsEventResourceRepository.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace CustomSerializer.Models
+namespace CustomEventSerializer.Models
 {
     public class AzureDevOpsEventResourceRepository
     {
