+++ samples/CustomEventSerializer/Program.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace CustomSerializer
+namespace CustomEventSerializer
 {
     public class Program
     {
